[CallingIdentityTokenDetector] Support methods with implicit receiver
Modified CallingIdentityTokenDetector to visit UCallExpressions instead
of UQualifiedReferenceExpression.

UQualifiedReferenceExpression has a structure of receiver.selector which
applies to method calls that use the class reference, however,
standalone methods will be ignored, i.e. method(). It is required to
consider standalone method calls because the class may be a child of
Binder and it can call its own method without any class reference.

Bug: 194085454
Test: atest AndroidFrameworkLintCheckerTest --host
Change-Id: I2b40afb747664cf4a36e441eaba1ecd64960e3b1

Former-commit-id: 0dee2ba23a07ffbf584da7a2945259ce7918e47b