Fix incomplete magnification mode transition
The magnification mode transition is interrupted by
the interaction client state changed. This accidental
invocation leads to consecutive transition mode requests.

To fix it, we check if the display is valid before ensuring
the windows on the display is tracking.

Besides, we also drop the new request when receving same
consecutive requests during the tranistion.

Bug: 194350828
Test: manual test
      atest MagnificationControllerTest
Change-Id: I9ef4314349b033971620a52dffbdf8df3085f351

Former-commit-id: 061238ca5f45f5534dd2cc79219bc8b2412e9010