Added device posture controller to SysUI.
This is mostly to add in the interfaces right now. This can be used to
notify surfaces about changes in the device posture when required. This
is currently implemented using the Jetpack device posture sidecar; this
needs to be used directly since androidx.window isn't available in
Gerrit right now.

Bug: 194280917
Test: Incoming.
Change-Id: I966fc5bdf43643bf570effbf6c8630abff5f86c4

Former-commit-id: 77eb288644b1db8dde526c85549811d8684cbea1