Fix TextureView docs to use more modern sample code
The existing TextureView class docs use a sample that relies on
Camera APIs, which have been deprecated for some time. The
new sample uses MediaPlayer (and calls out to other possible
components that could be used with TextureView). The docs have also
been updated to reflect recommendations on using CameraX for camera
functionality and SurfaceView vs. TextureView for performance reasons.

Test: N/A, docs only
Fixes: 191847306
Change-Id: I2038c09309ed63c79b583779124d48016ca225a2

Former-commit-id: 8e1a9dd5906603d267fd39a5183384b205824d38