Fix multiple control providers
Found an error in the logs:

ControlsBindingControllerImpl: Provider for
token:android.os.Binder@147a7f9 does not exist anymore

And then noticed that the ControlsActivity was being started by
framework when transitioning between ControlsFavoritingActivity and
ControlsProviderSelelctionActivity. This is strictly due to window
animations being enabled, and framework realizing that the
ControlsActivity will be visible momentarily between
activities. Previously uiController.show() was in the onStart()
method, which resulted in ControlsBindingControllerImpl rebinding to a
different service.

By moving to onResume(), the call to uiController.show() can be
avoided, hence avoiding the service rebinding so the controls can
correctly be shown in the favoriting activity.

Fixes: 191589943
Test: manual (requires multiple control services)

Change-Id: I634ab7c386c2098dc8f9971d320a8b80b8cd429a

Former-commit-id: 939a99cd4d8c536b935585444c2f2296fa563247