InputManagerService: Do not remove input channel when disposing monitor
To get around a race in InputDispatcher between the removal of the input
channel throught nativeRemoveInputChannel and the InputChannel#dispose
call, we remove the former call from InputManagerService.

This means it is now the responsibiity of the client to properly close
the input monitor by disposing of the InputChannel and all of its
duplicates.

Bug: 189135695
Test: manually try to reproduce the flaky conditions through a shell
script

Change-Id: Iba9f7c2428229ff6e2a7450a39828aed61160a3d

Former-commit-id: b3c3021841c9ddcdf51568f7ee1ab48d73abc8ff