Fix double-tap with non-accessibility focus behavior
we inject evetns to the location of hover events when satisifying
the following conditions
1. accessibility-focus node could be in any window or last touch window
   is an active window
2. no accessibility focus node

We reset last touch window id when user interaction is end. It leads to
failure if there is no service could interactive with windows.

To fix it, we don't reset the window id in such condition.

Bug: 194489345
Test: atest TouchExplorerTest
Change-Id: Ia0805eee1ca78254a7e73b63b19c4c2b2769effa

Former-commit-id: 075dd3432e0bb9f9b613b627c58622ea83e3e79a