Moving smartspace to the right on the lockscreen for split shade
There is a new view managed by ConstraintLayout guidelines that serves as container for smartspace and is placed on the right.
It's visible only in split shade and then it contains Smartspace. Smartspace view is switching between the default (current) view container and the new split shade container.
Switching Smartspace view and visibility of its containers is managed by LockscreenSmartspaceController.

Bug: 193654214
Test: KeyguardClockSwitchControllerTest, LockscreenSmartspaceControllerTest, KeyguardClockPositionAlgorithmTest
Change-Id: I1afcdf7f3c750c36b995f1d6fd9d4393eee8cbef

Former-commit-id: d04b85a03769f3eb68eda6aafa0d0291ae9c09b0