Fix ANR when switching the app task but IME selection dialog focused
As CL [1] that we don't give the IME child window focus when switching
to the next app task, we also need to handle IME selection dialog
focus and dismiss the IME selection dialog when the IME control target
will be changed in later.

To prevent the flickering when focusing to the next app task and see the
IME selection dialg dismising during switching the IME client, add
"imeParentChanged" parameter in
InputMethodManagerInternal#reportImeControl to ensure IME selection
dialog can be dissmissed before the IME control actually changed.

[1]: I72618745542f33bcbb1001bd3b812e79f7652dda

Bug: 194214768
Test: manual as issue steps:
    1. Launch an app to show soft-keyboard
    2. Tap IME icon from navBar to show "Select input method" dialog
    3. Start settings or other apps from SystemUI StatusBar
    4. Press back key wait for a monment, make sure ANR won't happen
Test: atest DisplayContentTests#\
        testImeMenuDialogFocusWhenImeLayeringTargetChanges

Change-Id: Iae3dd713b1e980067d28debea4e0a03707aa7938

Former-commit-id: bd868b666e95093fb1a92bffaeb8f9f757de444c