Use size compat bounds for window layout
When an app is in size compat mode, the configuration bounds
and layout bounds are not the same such as the app keeps
a larger configuration bounds and needs to be scaled when
drawing on screen. The size compat bounds are the scaled
bounds in screen coordinates so it should be used if available.

Fix: 194820928
Test: SizeCompatTests#testFixedScreenBoundsWhenDisplaySizeChanged
Change-Id: Ib97038560e65699f5275d6f5dc4b8e5881024e57

Former-commit-id: 9d4a077de92f44ab53e705f5f02ab9743db1d5a9