Annotate TextServicesManager with @UserHandleAware
This is a follow up CL to my CL [1], enabled TextServicesManager to
determine the target user ID by Context#getUser().

This CL aims to clarify what methods in TextServicesManager follow
this semantics.

Note that @UserHandleAware is still a kind of hint for readers.  There
should be no change at the runtime.

 [1]: I31dda3ae8795190d44b0622b8335c34ddbc5dd48
      ae8af708cd5b9ee8078822263593061c2e51f70a [formerly 9f141ee80230f266caa596a60e6d1278f072ac7c]

Bug: 192412909
Test: presubmit
Change-Id: I00bd99ef9d83a13bd60a5aefd22094725b67e271

Former-commit-id: db240e2687181fd89b91e865b160f1861c4d4ab5