Clarify how Handler is used in IC#getHandler()
Since its beginning [1] the system has been using

  InputConnection#getHandler()

only for retrieving Looper associated with it.  Although it has been
intentional, it'd be probably helpful for app developers if we can
clarify a bit more on this, because most likely the current behavior
is not going to change.

This CL also annotates the API with @Nullable as it should have been.

 [1]: Id9e579bb3e2966986cdcb1c34bc8cacfeca2e1a9
      98bb2e49ec06f3d303488a8cafe5427abadf804b [formerly 612cce92ad96eda1146c3abd2afa7aaa4d4f2b3f]

Bug: 26945674
Fix: 195434171
Test: presubmit
Change-Id: I14910c6f0cc43feec505b2951625d0c00203ac92

Former-commit-id: bebf16c4948fb8dd890690c548a1f373f75b8e0e