Finish existing secondary containers when starting to side
When a new activity is started to side of a primary container in an
existing split, finish the old secondary container by default to
replace it instead of putting on top.

Note that this rule only applies for new activity launches to side
via rules or dedicated APIs. If an activity is started regularly,
it would normally be placed in the existing secondary container on
top, not creating a new one and replacing.

Bug: 194996352
Test: Start to side from the primary container several times,
observe secondary container being replaced with subsequent launches.

Change-Id: If4439368a669c6252c97471f0e77286357285c9d

Former-commit-id: f45a4644090317132215b62d0d7bf0a1971f7c57