Accessibility for privacy status animations
This CL adds contentDescriptions to the privacy event chip and the
ongoing privacy dot views. This is a partial fix which allows talkback
for the ongoing privacy chip, and allows talkback scrolling to the
privacy dot. However, due to the fact that the dot lives in a different
window, the privacy dot shows up in the wrong part of the accesibility
tree and still needs to be virtually placed in the status bar.

Bug: 187197696
Test: manual
Change-Id: I6210e18c75a1aab61cbcf619a9a31c459fffe544
(cherry picked from commit 6da83090d88ce344bf1f0d97cc30ad039a59ae46 [formerly c14dd63160ee9fb303e3cd92ff63825a1303362a])

Former-commit-id: 421a14f3bce7c69040ac0482cae450f1bb68ba2d