Destroy ActivityRecord when it is removed from the hierarchy
When activity is finishing from back, it may be added to the
ActivityTaskSupervisor#mFinishingActivities and wait for the resume
activity IDLE before destroy.

However, TaskFragmentOrganizer will be notified to apply WCT to delete
the TaskFragment that has no running activity before that is done. It
will cause the activity removed from the hierarchy before the destroy is
called.

Fix: 195454840
Test: Verify with Sidecar + Demo app
Change-Id: I60d9217d061ee53421661bde496b5fb320f01b87

Former-commit-id: eccc88a8b5c628c99d569ec8bdf2d66b9373a747