Keeps the intercepted activity on the existing secondary container
When intercepting an activity start request, a new split pair
was always created and launch the activity to the side of the
caller activity.

This CL launches the activity on top of the caller activity
if the caller activity is already on the secondary container.

This is needed for the case that the activity is started
into a new Task, or new Task will be escaped from the current
host Task and be displayed in fullscreen.

However, the interception won't work if the activity is started
from another process.

Also adding logs for TaskFragment errors.

Bug: 194997356
Test: manually, start activity with sample app
Change-Id: I268c1dad191cdfdbf48cac84e18f1dfda761b0d3

Former-commit-id: 40fc6de5a7a44b86b459966a4208136c91fb0afb