Listener on system side & Return correct values
A listener is needed on the system side to make sure the
CompleteEconomicPolicy class is updated when any factor in the
JobSchedulerEconomicPolicy or AlarmManagerEconomicPolicy is updated by
the user. Both economic policy classes should also return the correct
values for their factors: the value inputted by the user or the default
value if the user has not put in a new value.

Bug: 196083166

Test: Visually check output of 'adb shell dumpsys tare'

Change-Id: I0dbef4cb23fcfefeeee444bdca26015840270d3e

Former-commit-id: c55d812c87738d68012d58bf042245e0ec71c5df