Allow to use fixed rotation if recents is animating
The condition of visible wallpaper was added for transition between
lockscreen and home that show wallpaper in different orientation.
But that also affects when swiping to home if the recents activity
is not home activity (3rd party launcher). And it will cause
additional rotation animation which looks unpolished.

This restores the case to have the same behavior as R.

Bug: 196012011
Test: Use 3rd party launcher as default home and gesture navigation.
      Swipe to home from an app which has different orientation from
      the launcher. There should be no rotation animation.
Change-Id: I8a3b4617434c619463a96b9e3106d25654a55391

Former-commit-id: faf59661282d73449593e5f240717b7361f8bed6