Apply fixed rotation on IME window if its target is rotated
This avoids flickering if the IME target activity and IME show in
different orientations. A new callback in WMS will be called if
IME window may be shown later, so the IME window can be told to
use the rotated configuration if it will show on a rotated activity.

Also consolidate multiple invocations from IMMS to WMS to a single
call onToggleImeRequested after requesting to show/hide IME. So it
doesn't need to acquire/release wm lock serveral times.

Bug: 195816362
Test: DisplayContentTests#testApplyTopFixedRotationTransform

Change-Id: I0b036594161898f9df6ad71f0adb5fe93880161a

Former-commit-id: 6b880493378abf01513b85bc2b00bdb471e3bfbf