Add haptic feedback constants for rotary-input haptic events.
These haptic patterns should be used when the rotary input on a device
is being used to scroll views. One provides some texture to the scroll
(e.g. light taps every few degrees), one provides a slightly heavier tap
when passing an element in, say, a recyclerview, and the last provides a
"thud" when getting to the end of a list.

All hidden for now; this is mostly to reserve the constants that are
being used by Wear. We can then expose these in a Compat library in the
future.

Test: Only new constants, no new functionality.
Change-Id: Ic85c21d4cb79184c2e5aa4a9f12d91a13db14e85

Former-commit-id: de210f4f64591e73f6820c0458f6cbc50af3b564