2/N Rename PluginInstanceManager to PluginActionManager.
Introduce PluginInstance.

PluginManager is in charge of coordinating the whole plugin system.

PluginActionManager is in charge of querying for the PackageManager
for plugins matching a specific action.

PluginActionManager creates PluginInstances for each package it finds,
and the PluginInstance is in charge of loading in the actual plugin
into the system.

This is another step down the path of being able to manually load in
plugins into the code before the PluginManager is ready.

Bug: 194781951
Test: atest SystemUITests && manual
Change-Id: Id525f27e362c29625168b56dc55f2899cd3f8810

Former-commit-id: ecc3306fdc7a988962984dd2cf09eabed673eaa2