Ensure that the keyguard container is always below message area.
In the current layout, the keyguard container is simply added to the
top-level NotificationShadeWindowView (descendant of FrameLayout) as a
sibling of the message area, so there's no way to align anything
underneath the message area (i.e. the "wrong PIN" bit). This is needed
for future devices, as new UX specs have the PIN entry field (the four
dots) immediately under the message area, rather than directly above the
bouncer.

Instead, this CL changes the keyguard inflation to inflate into a
FrameLayout directly underneath the message area. LinearLayout weights
are used to ensure it consumes all the remaining space.

An alternative would be, of course, to move the message area into the
keyguard views instead. This doesn't work, as the keyguard views are
lazily inflated (even keyguard_bouncer.xml), and the message area is
injected into some classes (via keyguard/dagger/KeyguardBouncerModule),
and thus needs to be available before it has been added to the layout.
It's probably fixable, but I really don't fancy rewriting half of the
Dagger injection right now.

Bug: 195012201
Test: Manually tested
Change-Id: I5d909d8548c3ae31ade92935b143a6cdcab46700

Former-commit-id: 6037800f125504e0e98eccd275d6b817227d37f3