Fixed an issue where the statusbar was made opaque even though it wasnt
Because on Pixel 4 and other devices, the light reveal scrim doesn't
have full alpha, the notification panel was marked opaque which removed
the wallpaper. This could lead to the wallpaper not being visible in
a couple of situations where it should have been, leaving AOD exposed.
Because of a separate bug (b/196306312), whenever the wallpaper is
hidden, we'd see a color layer, which would lead to the screen becoming
flickery and awake on AOD.

Fixes: 194760238
Test: atest SystemUITests
Change-Id: I5a7e80e2ced8f15f1b215549d0c051f1d61d2953

Former-commit-id: 5d993ac26c9aa87b1bd5783ccd261e7e7fcde26d