Only update the display when brightness and state are valid.
If either are invalid then we're in an indeterminant state and need to
wait until that's resolved. This should only happens at boot and then is
immediately corrected when we set our initial desired state. By not
checking the validity of the to-be-applied state, we're actually turning
the display to our assumed state briefly (i.e. OFF) even though we don't
know what state we actually want yet.

Fixes: 196566901
Fixes: 196274559
Fixes: 195071558
Test: Reboot device, no flicker in boot animation
Change-Id: Ie5a4c186e3af31eb8810f5c7014a4e2cc7cdbf91

Former-commit-id: a00cda677956e0553a6de4df8f25ebc9b2b3039e