Prevent repeatedly resuming/pausing activities in split
The embedding activity was repeatedly pausing/resuming
when the app is targeting SDK before Q. That's because
we were only comparing the z-order of the root task.
However, both of the two activities in split are in the
same task.

Update the topmost activity by comparing the z-order of
the activities.

Bug: 196923917
Test: start app with targetSDK prior Q in split
Change-Id: Ib6ffb12439689f38e83e320a0c9e739829e0f8fc

Former-commit-id: 0ae8f3504b96cb93586ed3f9634126d063505061