UserLifecycleTests: wait on storage before restart
When a user is stopped shortly after having started, the timing
results depend heavily on whether StorageManagerService had managed to
forge its connections yet. If it didn't, then stopping takes an
unrealistically short time. If it did, but is still in the process of
dealing with them, then stopping takes an unrealistically long time.

The solution is, for all appliable tests, to wait until storage is ready
before proceeding to stop or switch the user.

Also, some javadoc comments are added and clarified.

Bug: 169449457
Test: atest MultiUserPerfTests:android.multiuser.UserLifecycleTests
Change-Id: Ide848685637abac43488c8d89a3d6f6567c7deca

Former-commit-id: 1226ec462d46494396173a1376b7a617129a96fa