Exclude CommunalSurfaceView tap region from keyguard.
This changelist excludes the fullscreen tap region
(minus header and footer margin) when communal is
visible. This allows touches to pass through the
notification shade window to the CommunalSurfaceView
underneath.

Note that this currently has no effect on the
CommunalSurfaceView as it is on top of the z-order.
However, subsequent changes will move the
CommunalSurfaceView into the correct z-order.

Test: CommunalSurfaceViewControllerTest#testTapExclusion
Bug: 197026983
Change-Id: Ib80bdbb788e20756c18646833c23f844273af82c

Former-commit-id: a4ffa0ac9abe7ca63e4545f8cd8eefd01586d502