Rename hidden isStopped() method.
We introduced a hidden isStopped() for autofill to query the Activity
stop status, this is a final method. But the app that also has a
isStopped() method will break because the method cannot be overrided.

Try to rename the method to let it not generic to fix this app compat
issue. But this doesn't guarantee we don't have name collision in the
future.

Bug: 196929439
Test: manual. install app and the app doesn't crash.

Change-Id: I270066f77a69346b693b8329691d14fbcf66b1c6

Former-commit-id: b4b410f4a610de26dd37ebfab580db2dcc7dc3ea