Requests window magnifoication connection if any AccessibilityServiceConnection is able to control magnification
It is preliminary change for AccessibilityService to control window
magnifier.
The window magnification connectation needs a little time to be
established.
To make the connection ready before AccessibilityService requests to
control window magnifier, we request the connection earlier when any of
AccessibilityServiceConnection is able to control magnification.

Doc: go/a11yservice_control_magnification_in_t

Bug: 197191895
Test: atest AccessibilityManagerServiceTest
Change-Id: I90786aa1f94013a29b9713c69c76d2442b0cb866

Former-commit-id: 07167fa05fe6516ad6860503ae3cccbeb91953a7