RESTRICT AUTOMERGE: [MTE] Update manifest IDs to align with NDK.
The NDK is built out of internal *-release branches. Unfortunately, the
internal and external manifest attribute IDs don't currently line up.
As a result, NDK-built apps with memtagMode='[a]sync' won't turn on MTE
on AOSP images, but work in internal images.

We'd like NDK-built apps to work out of the box with AOSP images, and
so let's line up these IDs to fix the silent failure. This is especially
important as testing is mostly done using AOSP-on-QEMU.

RESTRICT AUTOMERGE because this should be an AOSP-only patch.

Bug: N/A
Test: Build the sanitizer test app
(https://github.com/google/sanitizers/tree/master/android) with an MTE
attribute and make sure you crash with MTE errors.

Change-Id: I48a3e08d01508dc6694fefd9563fb7d3f25108a9

Former-commit-id: 304bb210a204bde62e33ae5a1379bacceb5e3ac4