Use non finishing activity for remote animation opening target
If an existing task is moving to front by launching a trampoline
activity on top of the task, the finishing trampoline activity
should be skipped when choosing the remote animation target.
Otherwise there is no opening animation because the topmost activity
of the task may not have a window, i.e. RemoteAnimationController
skips the animation due to ActivityRecord#createRemoteAnimationTarget
returns null.

Now the mode of remote animation target is set explicitly when
requesting the animation, so the animatable container can choose the
proper target according to the mode.

Bug: 196636209
Test: atest RemoteAnimationControllerTest# \
            testOpeningTaskWithTopFinishingActivity
Change-Id: I0642dabec8889e101f831465a98707caba708362

Former-commit-id: 6f0cd1ecca8f839ac9d730dd509071ddd7a2b6d9