Make window metrics contains all insets information
The window metrics is mostly for the case the window is not in their
maximum bounds to get hints. If the window is not there yet, and the
display is in freeform, current implementation will assume the window is
in freeform windowing mode, and the system bars will be removed from the
insets. This will also happen to the windows with other windowing modes.

This behavior is not consistant with the API definiation and will cause
problem to ViewTest#testGetWindowVisibleDisplayFrame on freeform
display.

The change is to make sure we consider the window is in fullscreen
windowing mode when we try to get insets for window metrics.

Test: atest ViewTest
Test: atest DisplayPolicyTests DisplayPolicyLayoutTests
Test: atest WindowMetricsWindowContextTests WindowMetricsTest
Test: atest WindowMetricsActivityTests
Test: m checkbuild
Bug: 179454891
Change-Id: Iae264b567c161bdf4f3d0248c5f81de92fd47828

Former-commit-id: 8299a4c9ed070361d4690d5c3368c18b833d42b4