Remove ChooserTargetService flow entrypoints.
Despite the `USE_CHOOSER_TARGET_SERVICE_FOR_DIRECT_TARGETS`
flag being hard-coded to true and thus triggering the
deprecated flow, the effects of that flow (modulo potential
bugs) have been suppressed since the ChooserFlag was set to
false. That is, the logic removed in this CL was a (somewhat
misleading) no-op; the *intended* behavior was disabled,
but there were possible unintended consequences of the
vestigial code (which I believe were addressed as of
ag/15662048).

Test: presubmits
Change-Id: I0a2aa6e8bd796e41f6ebc69a54a1f74f9eeebbd9

Former-commit-id: 1e781efd02c0db80a0d0161bb001d0fc59e2897a