Use the height as width when there is a rotation needs to be applied for the dst rect.
Removed if condition check for dstRect and srcRect.

The AOSP camera for a split second sends the actual dstRect while
switching camera and that when the image scales because I wasn't doing
the conversion of the rect from width to height.
This scaling basically gives the perception that switch is slow, while
last few frames are basically still frames aka last buffer queue.

BUG: 197262426
Test: Verified manually on AOSP that camera switch is smooth.
Change-Id: Ic4b0b2ff3be2056a0c90841ed880e49c34ea2781

Former-commit-id: 98ccd61fa56bf9d45ea72e7f86afb3d91ce2dedb