[services.incremental] Use shared libutils for remount
Bug: http://b/197965342

services.incremental depends on libbinder.so, which already depends on
libutils.so.  When linking services.incremental with libutils.a, the
linker tries to open objects from libutils.a to redefine undefined
symbols in libbinder.so.  This causes a linker failure with upstream
LLD.  Switching to shared libutils.so instead doesn't increase the
dependency closure for services.incremental.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: I2735461ae64ef2e4c0afc170f7b765c1b5b9432e

Former-commit-id: cebf0493bfe008fe57ead09fa3f45a45ee134204