Remove ZigZagClassifer from lock-icon longpress falsing algo
If the touch leaves the lock icon area, we already drop the touch.
Therefore, we don't also need to also take into consideration the
ZigZagClassifier for the lock icon.

Also, only play the longpress vibration if we will be bringing
up the bouncer. If the FalsingManager think we're falsing, there's no reason to play
the longpress vibration.

Test: manually longpress lock icon on device with
udfps capability (after lockdown or reboot) and observe
longpress consistently brings up the bouncer
Fixes: 197271526

Change-Id: Ic9fa82a549599e48d281b4db69c04627803c6c5a

Former-commit-id: 374f321fca4572401911f1493122600140e918eb