Fix task missing after dropped an app in PIP to enter stage split
When dropping an app in PIP to enter stage split, the existing PIP task
will be reparented to the side stage root before PipController be
notified of the activity restart attempt, which result to the pinned
task didn't dismiss properly in wm-shell. This patch leaves pinned task
reparenting operations to task organizer to make sure PipController in
wm-shell can dismiss the pinned task properly with on activity restart
attempt callback.

Bug: 195727079
Test: Put Maps app in PIP, drag and drop the Maps icon to enter split,
      observed Maps expands to fullscreen and won't be put to
      background unexpectedly.
Change-Id: I8d8e4c0e86e4fc732d70d4a3d8e2c10701bc88d1

Former-commit-id: a75330fa8a9482ca0b44bbc214a4790aa81e282e