Don't show the embedded activities if visible needed activity without showWhenLocked
While the host activity and the embedded activity are displayed
side-by-side. Only the activity with showWhenLocked flag is shown
when the lock-screen is displayed. This CL updates the logic for
embedded activities policy if the visible needed activity without
showWhenLocked flags. Don’t display the embedded activities while
the lock-screen is shown.

Bug: 197298936
Test: atest TaskFragmentOrganizerTest
            TaskFragmentOrganizerPolicyTest
            SplitActivityLifecycleTest
Change-Id: I8e318c4b47f38ff3daf0713ab70cd980e7928000

Former-commit-id: 4f99b0dcac7ba8953fd66d6dd4185d28b4c06297