Normalize findPreferredActivityNotLocked()
Bug: 187080582

This change modifies PackageManagerService in preparation for moving
findPreferredActivityNotLocked() into Computer.  The synchronized
block in findPreferredActivityNotLocked() is pulled into its own
function to comply with Computer rules.  Two functions that are needed
by findPreferredActivityNotLocked() are made static so they can be
referenced from inside ComputerEngine.  Note that because the
functions can be made static, they do not depend on the PM snapshot.

Some calls from findPreferredActivityNotLocked() that used to be
outside the PM lock are now inside the PM lock.

A few legacy lint errors are corrected.

Test: atest
 * CtsContentTestCases:IntentFilterTest
 * CtsDynamicMimeHostTestCases
 * CtsRoleTestCases
 * FrameworksServicesTests:UserSystemPackageInstallerTest
 * FrameworksServicesTests:PackageManagerSettingsTests
 * FrameworksServicesTests:PackageManagerServiceTest
 * FrameworksServicesTests:AppsFilterTest
 * FrameworksServicesTests:PackageInstallerSessionTest
 * FrameworksServicesTests:ScanTests
 * UserLifecycleTests#startUser
 * UserLifecycleTests#stopUser
 * UserLifecycleTests#switchUser
 * FrameworksServicesTests:WatcherTest
 * android.appsecurity.cts.EphemeralTest
 * android.appsecurity.cts.InstantAppUserTest
Change-Id: Ib977ea97d06a5eaf02080272f25e8f6139406e7d

Former-commit-id: 37dc1932703b8200cbf4d370e7d480fd92dff8e6