[1/n] Letterbox Reachability: Repositioning on double tap.
This is the first CL for ensuring reachability of letterboxed portrait-only apps in fullscreen landscape device orientation.

The next CLs will add animation, restrict aspect ratio to split screen window size and add educational hints for the gesture.

Current CL allows to customize left and right positions of the letterboxed windows and initially opens the app on the right side of the screen.

Bug: 197549949
Test: atest WmTests:SizeCompatTests
manual with adb shell cmd window set-letterbox-style --isReachabilityEnabled true

Change-Id: I5a4f65510019c0dfc2b7d61efe0847d943467443

Former-commit-id: a75783bec4e9b20e6549faf6f53a852176811782