[SettingsProvider] additional logging for debug telephony bug
For the past few months we have been seening a telephony bug that
happens occasionally without a reliable way to reproduce it. The bug is
that for some reason, all of a sudden, the telephony thinks that the
device is unprovisioned, even though the device is provisioned. Based on
initial inspection, the inconsistency is from Settings Provider.
However, the dumpsys shows that the DEVICE_PROVISIONED value is 1, and
no other clients have reported this issue. The inconsistency could be
from memory corruption, though that would probably have been reported by
other clients as well.

To further assist debugging, I am adding additional logging just for the
query on DEVICE_PROVISIONED. This will at least show what exactly is
the value returned to telephony, when the bug occurs again. I will
remove the additional logging as soon as we have some clue.

BUG: 197879371
Test: builds
Change-Id: I1321168081fa808dca4503b84d58983df145e790

Former-commit-id: 9ee3eebb99a67200979ca073dbbac8adba0a100d