Manually register Dumpables to DumpManager
Dependency.get() automatically registers its managed objects to the
DumpManager if they implement Dumpable. However, if code is refactored
that removes all usages of Dependency.get() in favor of Dagger
injection, then there is no longer a guarantee that the class in
question will be registered to DumpManager (and so included in bug
reports).

Instead, remove the auto-registration feature of Dependency in favor of
manually registering all dumpables with the DumpManager.

Bug: 198713580
Test: atest
Change-Id: Ie02a44fb7da0b76bf53da874cc9eee030a1b9173

Former-commit-id: 3aca431b69dcf12a531ef522a183f552f49e9866