Don't hold hibernation lock when calling other services
Fix deadlock issues by ensuring that calls to activity manager and
package manager are done without holding the hibernation lock.

This will loosen some guarantees on the hibernation state (e.g. there
may be a brief period of time where an app is hibernated but not
force-stopped), but none of the callers currently need that strong
guarantee.

Bug: 196535754
Test: atest AppHibernationIntegrationTest
Test: atest AppHibernationServiceTest
Test: adb shell cmd jobscheduler run -u 0 -f
com.google.android.permissioncontroller 2

Change-Id: Ic67d9a93565a935c077707e3c34a4c6605f041d1

Former-commit-id: d41db18129f4119218c6eda21a7d51c0313b18c1