Avoid applying window config from updating persistent config
The changed global configuration will also apply to default
display. When the display is rotating, the transition will be
disturbed if the given configuration contains stale window
configuration (e.g. bounds and rotation).

Because persistent configuration only includes locale and font
scale, clear the window configuration to avoid the noise.

Bug: 196313047
Test: Back from a landscape activity to a portrait activity.
      The portrait app activity calls LocalePicker#updateLocale
      in onResume. The transition should not be flickering.
Change-Id: Ib1bea1b752eb695e26b8b0622bcf06ec01cce3e6

Former-commit-id: b4fdc0e4ba0796eebb215be06782252d4110cce7