Fix flicker when task move to back
When task move to back, the task visibility will be invisible
temporarily because visile app task reparent to TDA and below task
still not move to front.

Because this visibility change, we will hide divider and stage root
task then it cause flicker.

Fix this by only update divider and stage visibility when both stage
visible or invisible.

Bug: 199369390
Test: manual
Test: pass existing tests
Change-Id: Ide0b277e26e03ea3d7fb20132ff0e21807eee332

Former-commit-id: 3ad28efb880fe9126603cb952a2fcdd8affa8b04