Limiting activity embedding to the activities of the same app
Disallow having cross-app activities being embedded until we
have a better understanding and solutions to deal with potential
security issues.

Also not allow embedding tasks even belongs to the same app
because the  activities of another app could still be launched
on the embedded task afterward.

Only system is allowed to embed tasks.

Bug: 189386461
Bug: 197364677
Test: atest TaskFragmentOrganizerPolicyTest

Change-Id: Ib6ca8783449872f035fffe51743c5a966768c894

Former-commit-id: d2ceb2ba6105c6137c27f92f03e874053e6d12f6