Stop lying about MissingMethodFlags.REQUEST_CURSOR_UPDATES
This is a follow up CL to my previous CL [1], which used
MissingMethodFlags mechanism [2] to let that API fail when the IME and
the target app are not running on the same display.

As discussed in Bug 194110780, however, there is a problem in the
MissingMethodFlags mechanism that it could have caused unexpected task
reordarings from the IME developer when sync InputConnection APIs
immediatelly fail without waiting for those tasks to be scheduled to
the target app then handled.

As a preparation to fully deprecate MissingMethodFlags, this CL
enables InputMethodManagerService to stop lying about

  MissingMethodFlags.REQUEST_CURSOR_UPDATES

when the IME is rendered on a diffirent display than the target app.
Such check is now explicitly implemented in

  RemoteInputConnectionImpl.

Other than that, there is no observable behavior change.

 [1]: Ie2f7a5117cff3a13ad5c5806fd4b3abef7569549
      57168d1de8e8354a9c04ba9388fc200abe7b1b27 [formerly 3d2cc0fffd13f46923d0fcfdd73d375ebbe955ce]
 [2]: I3c58fadd924fad72cb984f0c23d3099fd0295c64
      4b1e09c18d4c13c9c1bee4f2ecbf5bbba575dd75 [formerly 19a80a1e807acd00bec999eaac7812da6ffce954]

Bug: 131368625
Bug: 194110780
Test: atest CtsInputMethodTestCases
Test: atest MultiDisplaySystemDecorationTests#testCrossDisplayBasicImeOperations
Change-Id: Iec12733d37e112b7271436bba15094ae2a55a450

Former-commit-id: 6d2e9786b47650b1e74729f5609ebe0982c2f86c