Make DreamActivity translucent when waking up dream
The DreamService calls onWakeUp, which allows the dream application to
finish the dream itself and apply an exit animation for up to 5 seconds
(after that, the system force removes the dream window). During the time
when the dream application is performing an exit animation, the dream
activity window is opaque. This causes problems when the dream
application applies a long exit animation - turnScreenOn+noHistory
activities are put under the dream (because dream has highest z-order)
and get destroyed before they get a chance to be drawn (because they
never become visible).

This CL makes the dream activity translucent before onWakeUp is called.
This allows turnScreenOn+noHistory activities to be resumed underneath
the dream, while the dream is transitioning out.

Bug: 162880999
Bug: 198359939
Test: manual
Change-Id: I5045ba8765dd1918cf85ec0fa433617843e727bd

Former-commit-id: 7b024ebd9d19489f6bc5ab8db07e46059555e50e