Consider the source task while evaluating the launch-root-task
e7538ef655 [formerly 87a79d3a1d], always launches in the adjacent task of the launch root
if the adjacent task is focused.  The behavior resulting in Chrome
launcher can not launch new tasks to the specified launch root
anymore because the adjacent task is always in focus. We should
think of using the source task to evaluate the launch-root-task
instead of the focus state.

Bug: 180079028
Bug: 198810232
Test: atest MultiWindowTests, WMShellUnitTests,TaskDisplayAreaTests
Change-Id: I2ad7da94412b1dff1e0deb631b8a2d9b9565eac9

Former-commit-id: d666f5462173ee2176cf1ece5698960db78e1ca8