Round 3: Fix flakiness of testNativeCrash
Round 2: ag/15616233
Round 1: aosp/1702009

Even with increased timeouts, testNativeCrash is not always able to get
the tombstone of the crashed native app. This is because the tombstone
timestamp can be more than 1000ms before the ApplicationExitInfo,
causing them to not match. See b/191798394 for more details.

This CL increases the maximum timestamp difference to ensure that the
tombstone (almost?) always matches.

Bug: 192299487
Test: atest android.app.cts.ActivityManagerAppExitInfoTest
Change-Id: Ideeb93dbcb40ce9d564f159fe1f7574554c09a40

Former-commit-id: 580909db4904aeedf10d4cf1a05d783e8d963df6