Flag whether a RemoteAnimationTarget has parent in animation
There is case when a TaskFragment is resizing while one of its children
is open/close. In such case, both the TaskFragment and its child will be
included in the transition, but we don't want to animate both of them.
Use the new flag to determine if the child need to be animated.

Bug: 196173550
Test: test with demo app
Change-Id: Ib5ad32501e046e5b612f6d5ae2faed6266d99f3d

Former-commit-id: 79ab7ea034096d868d58f2c9a3f5ee9f50510177