Log a notification update when fields such as importance or ranking score are updated.
This update happens in the function performing the actual sort because while adjustments are queued, the actual updates to fields in the NotificationRecord don't happen until sorting time. To avoid spurious extra logging, we only log an update to notifications that were updated via assistant adjustments (by tracking those keys in a set).

Note that there are two ways adjustments can be applied to notifications: synchronously to an enqueued notification or asynchronously (prompting the sort). Given that enqueued notification (posts or updates) are logged once the post or update goes through, this change only logs the updates to the ones updated asynchronously. (Otherwise, we may see a notification "update" before the notification is ever even visible to the user.)

Bug: 195579280
Test: manual via statsd_testdrive, atest NotificationManagerServiceTest
Change-Id: I97ef8865e31653f0fee0ac943251fecc21e5e05f

Former-commit-id: a92ad322d1a9a9098c1f2f21817f695f338a0570