Clean up the no-history entries while the app died
mNoHistoryActivities, A list maintains the no-history activities.
If the device goes to sleep with such activity in the paused state
then save it here and finish it later if another activity replaces
it. We should clean up the entries together while the app dies
otherwise the items still stay in the list.

Bug: 200159293
Bug: 199523354
Test: 1. start some activity with no-history and transparent property
      2. adb shell am force-stop “the app”
      3. adb dumpsys activity activities to check mNoHistoryActivities.
Change-Id: Ib2e3f12f83d562cd25efe918eb27e25f4dc890b2

Former-commit-id: a6659239288f281e61bf5fb2d12633a1be6f54f2