Increase Bluetooth MIDI maximum size
The current Bluetooth MIDI max size is 20, so certain keyboard
creators like Yamaha are complaining at the number of packets
sent at a time. This cl changes the max size to 512 bytes.

Previously, a 2D array of size 20 was created. With this change,
this 2D array is no longer created. Instead, the last array is
cached. Only when a differently-sized packet is created, is new
called on the array.

Bug: 147950483
Test: Connected to Alesis drums via bluetooth and tested both input and
output. Also verified that BluetoothGatt: configureMTU() was called with
an MTU of 512 and the omMtuChanged callback returned a valid status

Change-Id: Ia10b81ffe51277d0565631fd28e71c7c8ca884ae

Former-commit-id: 3688b228ce54cbd3b3e385df44462fc17ab01fb4