Only allow completion callback to be run once for PIs
PendingIntent.send() should either throw an exception OR run it's
completion callback, but not both. However, a bug that is not fixable in
the near term means that it's possible for both to happen. In order to
avoid the completion callback from being run even if send() fails, we
gate it.

Bug: 199464864
Test: manual + presubmits
Change-Id: I993436b2bd2ec5240be92e6d3815b3a9edef970f

Former-commit-id: 0ec832bba2ec5c5e45ff03184ecb95890447acb2