Fix test failure on WindowMagnificationControllerTest
In the failed test, we try to keep the same position after
the rotation. The center of the window is at the center of the
screen by default. We assume the screen is rectangle however
it seems something wrong on the emulator so the center X
and center Y are the same, hence the window doesn't have any offset
after the rotation.

To fix it, we make the center X and cetner Y different.

Bug: 200799671
Test: atest WindowMagnificationControllerTest on cf_x86_64
Change-Id: Id0482359b4b1c5c26f3e3f3f468126ae523870f7

Former-commit-id: 4197e17a5796c838eca86cb473b5e96a3180b67c