Populate PointerCoords values using values from MotionEvent
Some values in MotionEvent, such as AXIS_X and AXIS_Y, will be
transformed into a different coordinate space. This means the value will
differ from the one stored in the native PointerCoords.

When populating the java PointerCoords object, we need to make sure to
use the transformed values from MotionEvent instead of the raw values
stored in the native PointerCoords.

Bug: 179274888
Test: atest MotionEventTest (CTS)
Change-Id: Iffa459fdeba1ef23fe3bf7e36e897d120d9c457c

Former-commit-id: 9dabf6bac519bde9432575195859810dadd9f74e