Do not change launch mode to FF is FS explicitly requested
When a non-resizable activity is launched in a freeform display,
it is launched in freeform windowing mode when its orientation
is different than that of its parent. This always happens even
if the activity requested to be launched in fullscreen mode
explicitly.

This CL changes the launch parameters so that the windowing mode
is only changed to freeform if it was not explicitly requested
that it be fullscreen.

Bug: 200058556
Test: atest CompatChangeTests
Change-Id: I0ad673a6a93c8e1cb1787d12feb49124f7d9c171

Former-commit-id: fe7c92ec40108db3d5e6bcd9a4ddf3b53e0f0292