Use PlatformProperties to read VelocityTracker strategy
The VelocityTracker strategy is now going to be accessed from
PlatformProperties. Update the usages in VelocityTracker.

Also, dump the input-related properties in InputManager.

Ideally, we would create a VelocityTracker and dump the strategy of
the default-created VelocityTracker. However, that would require the
following additional work:

1) Moving the VelocityTracker strategy initialization logic to native
2) Adding VelocityTracker to NativeInputManager's dump

To avoid this additional work, just dump the system properties for now.

Bug: 192439431
Test: run 'adb shell dumpsys input' and observe the output
Output:
```
Input properties:
  per_window_input_rotation = false
  persist.input.velocitytracker.strategy = impulse

Input Manager State:
```

Change-Id: I7594973f28abaf75edd39d03c952b5a057779a45

Former-commit-id: 84b38799cd4f44d79ecfdcaf7a5de35e0edb38e7