Correct launch task behind when dream is present.
Logic was introduced to launch any activity behind
the current task when a dream was present. This is
problematic when a dream is stopping as the top
task will not be the dream and the new activity will
be launched behind the task previously behind the
dream.

This changelist addresses this corner case by
checking the stopped state of the DreamActivity before
deciding to place the new activity behind.

Test: CtsWindowManagerDeviceTestCases
Bug: 200752015
Change-Id: I050eac31e1a40c885b24565f5a4c3bb1a053c2a5

Former-commit-id: 74e50739dd595c9c3d676322d1a356a7b65e965f