Check pending rotation change after recents animation
Rotation change was skipped while running recents animation to
avoid interrupting the animation and may apply unexpected
intermediate orientation to recents activity.

But the rotation change should be still applied after the animation
is finished. Otherwise the display may stay in old rotation. The
case may be more obvious since live tile is introduced, because
the animation keeps running until switch to app. E.g. enable auto
rotation and the top is a rotatable app in landscape, and then enter
recents and rotate the device to portrait. After going back to the
current app, the display is still in landscape.

Fix: 199371058
Test: RecentsAnimationControllerTest#testCheckRotationAfterCleanup

Change-Id: I5e7487b08e139b251d160bd0627326b83351640e

Former-commit-id: 36fd1663024b3c39d25b6758a789daa046ff998a