Applying lifecycle effect when changing TaskFragment bounds
The activity visibility was not updated when folding the
Device and the TaskFragment below remain visible and the
activity was still in resumed state.

The same issue can also result in wrong activity visibility
state when unfolding the device due to there is a race between
processing the activity configuration changes and requesting
TaskFragment bounds changes from the organizer.

Bug: 201564759
Test: resize display when activities in split
Change-Id: I2bffb2ee27fc2437be231df606e2e46509dced97

Former-commit-id: 6f8ebc05d9f62ab6aea0535bbf1f8c2df9ac7a9f