Introduce UM.getProcessUserId
UM is being made @UserHandleAware, so using an explicit method avoids ambiguity.

Ideally, we would just use UserHandle.myUserId(), with no need to go via
UserManager. However, this method is static and is therefore very
difficult to mock in all the various places that mock it, even with
extended mockito. It's much much simpler to simply switch the name of
the hidden API.

Bug: 201679968

Test: (No-op) compiles and treehugger
Change-Id: If93eda899b52d0acae995ec96f2491f5ea1a992d

Former-commit-id: 953a49b58133e558154588da84363b1122c381e0