Use display awake signal to replace screen-off signal for sleep token.
For sleep token, device goingToSleep/awake might be more reliable than
screen on/off since it can be affected by AOD, we can consider this as
a signal to sleep device.
And move the SleepTokenAcquirer to DisplayPolicy, we don't need the
SleepTokenAcquirer interface on ActivityTaskManagerInternal anymore.
Also reuse the acquirer on RootWindowContainer for default display
so there does not need to create another one.

Bug: 199633495
Test: atest KeyguardTests KeyguardTransitionTests
Change-Id: I676cd5855f31d31d7acba9f161d27bdaf58844f1

Former-commit-id: 0760ca28ea0652639954003564b3e1d35fe8d324