Do not set profile owner on guest users on headless system user.
The guest user cannot have extra apps, hence the DPC app is not
installed. We could try to bypass that restriction, but it could
cause other issues down the road; besides, on phones the guest
user is not managed neither and DPM doesn't provide an API to
created managed guests.

Test: manual verification using TestDpc
Fixes: 202327386

Change-Id: I02f963bc84325dd73a18abb7103fc34d4960005d

Former-commit-id: ec902a9694969e1ea9f9298e11a193a30aade797