Prevent prefetch jobs from having deadlines.
Now that we're trying to run prefetch jobs close to app launch, there's
no good reason for a prefetch job to have a deadline. Periodic jobs
don't run until the constraints are satisfied. Just wanting to run a
prefetch job by some time just for the sake of running it is not
something we want to allow. Gated by target SDK, updated apps will be
prevented from scheduling a prefetch job with a deadline.

Bug: 194532703
Test: adb shell am compat disable 194532703 android.jobscheduler.cts; atest android.jobscheduler.cts.JobInfoTest#testPrefetch
Test: adb shell am compat reset 194532703 android.jobscheduler.cts; atest android.jobscheduler.cts.JobInfoTest#testPrefetch
Change-Id: I92d91c89d01765f76064dd00725137f166ca483f

Former-commit-id: 85033f3306455108ecb96c149c74cbc65de986bf