Add log when upfps overlay isn't in sync with KeyguardUpdateMonitor
Log a debug msg when UdfpsController hides the udfps overlay when
KeyguardUpdateMonitor thinks that fingerprint is running. If we
get to this log, we expect a KeyguardUpdateMonitor fingerprintRunningState
update to 0 shortly afterwards. If not, there's a discrepancy between
UdfpsController and KeyguardUpdateMonitor which can lead to a bad state.
For example: the udfps icon not showing, but KeyguardUpdateMonitor thinks
that it's showing so UDFPS will not be re-requested.

Test: manual
Fixes: 197167446
Change-Id: I4cc97ae0fcc2f11a320fdcd014fbcf14227a011c

Former-commit-id: 0142b7345fe6aa82189e3cf8ae2dbdb8756a0513