Support system app update leaving shared UID then getting uninstalled
It is possible for a system app to leave shared user ID by an update. We
need to keep track of this in case the upgrade is uninstalled. In this
scenario, the shadowed system package will be reinstalled using the
original shared UID. This has the side effect of the app ID used by the
shared user group being held forever even if no apps is actively using
it, because only the app ID is stored in packages.xml, not the
sharedUserId name, and we have to use the app ID to link the disabled
PackageSetting with the corresponding SharedUserSetting.

Test: tracking with b/202993696
Bug: 179284822
Fixes: 201509122
Change-Id: If48aeb7d63daf1ef8bdf06a9d2dd34c1de8603c8

Former-commit-id: d8597aa9a5bc8def9520bb78bb6e9b368292cbdc