Ensure that ghosted view is attached when calculating its matrix.
This CL ensures that the ghosted view is attached to a ViewGroup before
calculating the matrix of that view. It could happen that the view would
be detached after the animation was started, which would then crash
SysUI when calling GhostView.calculateMatrix.

I tested this CL manually by first reproducing this b/202668591 by
detaching ghostedView from its parent after the animation started, and
then confirmed that this fix prevents the animation from crashing.

Test: Manual
Bug: 202668591
Fixes: 202668591
Change-Id: I7bf0bedb4860a00f1c44cac9258ff5dc753fb355

Former-commit-id: d2dc31332a89a6f9eff46f6fc2353e9861cf9607