Do not animate statusbar unfold when in portrait orientation
Disables unfold animation for statusbar when it is
perpendicular to the screen fold.
Moves ScopedUnfoldTransitionProgressProvider from launcher
to shared SysUI/Launcher code.
Introduces NaturalRotationUnfoldProgressProvider which emits
transition events only when in natural rotation.
Added proper disposal of resources when status bar is detached
as PhoneStatusBarView is re-created when folding/unfolding.

Bug: 201518277
Test: manual
Test: atest com.android.systemui.unfold.util.NaturalRotationUnfoldProgressProviderTest
Change-Id: If4da3afca2b41c3675ccf6838e9ad6dac1b6f82e

Former-commit-id: fa92428107ba7fccd591870234a80e5c87ea1200