Migrate are/setNotificationsEnabled to PM
To test this I'm temporarily created a new NMS level test file that
is identical to NMSTest's configuration except that migration is enabled.
Since the setting has to be set before NotificationManagerService is enabled,
it's hard to set on a test method level whether the migration should be enabled
but it's easy to set on a test class level.

Test: atest
Bug: 194833441
Change-Id: I476ee85c06e430a1c6305e53d17103eb3128ae31

Former-commit-id: 1184a696157b6bc7952666632606b9e98f9d1625