Attempt 2 for adding OWNERS for SNTP code
Commit 9d024fb0d7848bd2a2449123d4f04a2aac0db637 [formerly e5ad1297165888eb5eef24c1985e64b18b8fb88b] combined with pending
reviews demonstrated the OWNERS syntax has quirks and it is not working
as intended.

Documentation used:
https://android-review.googlesource.com/plugins/code-owners/Documentation/backend-find-owners.html#syntax

The syntax guide recommends that patterns like sub-dir/* aren't used as
they can cause problems when renaming directories; better to have OWNERS
in the sub-dir. The pattern sntp/* didn't appear to be working (perhaps
because of spaces).

Test: None
Bug: 199481251
Change-Id: I193744eaaad529b9680d3ccf905b86d68f62f7ff

Former-commit-id: 3725e9c63814e9cf00b7f08d68f5c7d311da3124