Remove legacy calls to StatusBarWindowView for layouts
QuickStatusBarHeader, KeyguardStatusBarView, and PhoneStatusBarView all
had a few calls to some methods in StatusBarWindowView that would try to
calculate the insets necessary draw content in the status bar area that
avoids cutouts. This CL removes those methods and moves that
functionality into StatusBarContentInsetsProvider so that everything can
benefit from the caching it provides and hopefully this simplifies
things.

Test: atest SystemUITests; manual
Bug: 203223072
Change-Id: I6235394a3c71a360b5cc1fb9f511a52aef8c8f17

Former-commit-id: 70a1745ebb00ef0de59d66d2b4803bc5459cdc26